--- src/main/java/com/fastcampus/jober/core/auth/jwt/JwtAuthenticationFilter.java
@@ -1,70 +1,0 @@
-package com.fastcampus.jober.core.auth.jwt;
-
-import com.fastcampus.jober.core.auth.session.MemberDetails;
-import com.fastcampus.jober.domain.member.Member;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-
-import java.io.IOException;
-
-@Slf4j
-public class JwtAuthenticationFilter extends BasicAuthenticationFilter {
-
-    private JwtTokenProvider jwtTokenProvider;
-
-    public JwtAuthenticationFilter(AuthenticationManager authenticationManager) {
-        super(authenticationManager);
-        if (jwtTokenProvider == null) this.jwtTokenProvider = new JwtTokenProvider();
-    }
-
-    @Override
-    protected void doFilterInternal(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain chain
-    ) throws IOException, ServletException {
-
-        String prefixJwt = request.getHeader(JwtTokenProvider.HEADER);
-
-        if (prefixJwt == null) {
-            chain.doFilter(request, response);
-            return;
-        }
-
-        String jwt = prefixJwt.replace(JwtTokenProvider.TOKEN_PREFIX, "");
-
-        if(JwtTokenProvider.isExpired(jwt)) {
-            log.error("토큰이 만료되었습니다.");
-            chain.doFilter(request,response);
-        }
-
-        Long id = JwtTokenProvider.getMemberIdFromToken(jwt);
-
-        Member member = Member.builder()
-                .id(id)
-                .build();
-
-        MemberDetails memberDetails = new MemberDetails(member);
-
-        UsernamePasswordAuthenticationToken authenticationToken =
-                new UsernamePasswordAuthenticationToken(
-                        memberDetails,
-                        memberDetails.getPassword(),
-                        memberDetails.getAuthorities()
-                );
-        authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-        SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-
-        chain.doFilter(request, response);
-    }
-
-
-}
