--- src/main/java/com/fastcampus/jober/core/dto/ValidDTO.java
@@ -1,13 +1,0 @@
-package com.fastcampus.jober.core.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class ValidDTO {
-    private String key;
-    private String value;
-}
