--- src/main/java/com/fastcampus/jober/core/exception/Exception401.java
@@ -1,23 +1,0 @@
-package com.fastcampus.jober.core.exception;
-
-
-import com.fastcampus.jober.core.dto.ResponseDTO;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-
-// 인증 안됨
-@Getter
-public class Exception401 extends RuntimeException {
-    public Exception401(String message) {
-        super(message);
-    }
-
-    public ResponseDTO<?> body(){
-        return new ResponseDTO<>(HttpStatus.UNAUTHORIZED, HttpStatus.UNAUTHORIZED.getReasonPhrase(), getMessage());
-    }
-
-    public HttpStatus status(){
-        return HttpStatus.UNAUTHORIZED;
-    }
-}
