--- src/main/java/com/fastcampus/jober/domain/member/Member.java
@@ -1,53 +1,0 @@
-package com.fastcampus.jober.domain.member;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "member")
-@Data
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false, length = 30)
-    private String email;
-
-    @Column(nullable = false, length = 100)
-    private String password;
-
-    @Column(nullable = false, length = 10)
-    private String username;
-
-    @Column(nullable = false)
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-    @Column
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    private LocalDateTime deletedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        this.createdAt = LocalDateTime.now();
-    }
-
-    protected void onUpdate() { // 회원정보 수정 시 호출
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    protected void onDelete() { // 회원 삭제 시 호출
-        this.deletedAt = LocalDateTime.now();
-    }
-}
