+++ src/main/java/com/fastcampus/jober/global/auth/session/MemberDetails.java
@@ -1,0 +1,54 @@
+package com.fastcampus.jober.global.auth.session;
+
+import com.fastcampus.jober.domain.member.domain.Member;
+import java.util.Collection;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Getter
+@RequiredArgsConstructor
+public class MemberDetails implements UserDetails {
+
+    private final Member member;
+
+    // TODO - 권한 찾는 부분.
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+//        Collection<GrantedAuthority> collector = new ArrayList<>();
+//        collector.add(() -> member.getRole().name());
+//        return collector;
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return member.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return member.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
