+++ src/main/java/com/fastcampus/jober/global/exception/Exception500.java
@@ -1,0 +1,24 @@
+package com.fastcampus.jober.global.exception;
+
+import com.fastcampus.jober.global.utils.api.dto.ResponseDTO;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+
+@Getter
+public class Exception500 extends RuntimeException {
+
+    public Exception500(String message) {
+        super(message);
+    }
+
+    public ResponseDTO<?> body() {
+        return new ResponseDTO<>(HttpStatus.INTERNAL_SERVER_ERROR,
+            HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase(), getMessage());
+    }
+
+    public HttpStatus status() {
+        return HttpStatus.INTERNAL_SERVER_ERROR;
+    }
+
+}
