+++ src/main/java/com/fastcampus/jober/global/utils/api/dto/ValidDTO.java
@@ -1,0 +1,14 @@
+package com.fastcampus.jober.global.utils.api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class ValidDTO {
+
+    private String key;
+    private String value;
+}
